consolidated sync + async filter classes
