+++ src/dpr/all.mjs
@@ -6,7 +6,7 @@ import CardGroup from './components/card-group/clientClass.mjs'
 import Filters from './components/filters/clientClass.mjs'
 import Columns from './components/columns/clientClass.mjs'
 import ToggleButton from './components/toggle-button/clientClass.mjs'
-import AsyncFilters from './components/async-filters/clientClass.mjs'
+import AsyncFilters from './components/async-filters-form/clientClass.mjs'
 import SyncFilters from './components/sync-filters/clientClass.mjs'
 import Pagination from './components/pagination/clientClass.mjs'
 import ReportActions from './components/report-actions/clientClass.mjs'
