--- src/dpr/components/async-filter-inputs/view.njk
@@ -1,11 +1,0 @@
-{% from "dpr/components/filter-input/view.njk" import dprFilterInput %}
-
-{% macro dprAsyncFilterInputs(filters, prefix) %}
-
-  <div class="filter-container">
-    {%- for filter in filters %}
-      {{ dprFilterInput(filter, prefix) }}
-    {%- endfor %}
-  </div>
-
-{% endmacro %}
