+++ src/dpr/components/sync-filters/view.njk
@@ -1,39 +1,10 @@
 {% from "dpr/components/filter-input/view.njk" import dprFilterInput %}
+{% from "dpr/components/filters-form/view.njk" import dprInlineFiltersForm %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 
 {% macro dprSyncFilters(filters, prefix = 'filters.') %}
   <div data-dpr-module="sync-filters" class="dpr-filters sync-filters-form--hidden">
-    <form id="sync-filters-form">
-
-      <div id="query-error-summary" class="query-error-summary query-error-summary--hidden">
-        {{ govukErrorSummary({
-          titleText: "There is a problem",
-          errorList: []
-        }) }}
-      </div>
-
-      <div class="filter-container">
-        {%- for filter in filters %}
-          {{ dprFilterInput(filter, 'filters.') }}
-        {%- endfor %}
-      </div>
-
-      <div class="filter-actions">
-        <div class="govuk-button-group">
-
-          {{ govukButton({
-            text: "Apply Filters",
-            href: '#',
-            attributes: {
-              id: 'sync-apply-filters-button'
-            }
-          }) }}
-
-          <a id="sync-reset-filters-button" class="govuk-link govuk-link--no-visited-state dpr-reset-columns-button" href="#" >Reset filters</a>
-
-        </div>
-      </div>
-    </form>
+    {{ dprInlineFiltersForm(filters, prefix) }}
   </div>
 {% endmacro %}
