+++ src/dpr/utils/RequestReportUtils.ts
@@ -4,11 +4,11 @@ import Dict = NodeJS.Dict
 import { AsyncReportUtilsParams, ExecutionData, RequestDataResult } from '../types/AsyncReportUtils'
 import type ReportingService from '../services/reportingService'
 import { ReportType, RequestFormData, RequestStatus } from '../types/UserReports'
-import filtersHelper from '../components/async-filters/utils'
+import filtersHelper from '../components/async-filters-form/utils'
 import { components } from '../types/api'
 import { DashboardDefinition } from '../types/Dashboards'
 import { Services } from '../types/Services'
-import { SetQueryFromFiltersResult } from '../components/async-filters/types'
+import { SetQueryFromFiltersResult } from '../components/async-filters-form/types'
 import type DashboardService from '../services/dashboardService'
 import { removeDuplicates } from './reportStoreHelper'
 import UserStoreItemBuilder from './UserStoreItemBuilder'
